Set CPACK_GENERATOR on MacOS
On MacOS, set "CPACK_GENERATOR" variable to "Bundle". That allows a
MacOS bundle to be generated with a simple

   % make package

command on that operating system. Also set the CPack variables only on
MacOS since CPack currently is only used to generate the MacOS bundle.
The source distribution is generated via the export/archive mechanism of
the version control system.